chore: 공통적인 로직 모듈로 분리
