+++ functions/removeErrorSpan.js
@@ -1,0 +1,8 @@
+const removeErrorSpan = (inputBox, input) => {
+  if (inputBox.lastElementChild.tagName === "SPAN") {
+    inputBox.lastElementChild.remove();
+    input.classList.remove("error-box");
+  }
+};
+
+export default removeErrorSpan;
