--- index.js
@@ -1,2 +1,0 @@
-import convertEmailErrorMessage from "./email.js";
-import convertPasswordErrorMessage from "./password.js";
