+++ signin.html
@@ -53,8 +53,6 @@
         </div>
       </main>
     </div>
-    <script type="module" src="index.js"></script>
     <script type="module" src="signin.js"></script>
-    <!-- TODO  나중에 합치기 -->
   </body>
 </html>
