+++ signup.html
@@ -57,9 +57,6 @@
         </div>
       </main>
     </div>
-    <script type="module" src="index.js"></script>
-    <script type="module" src="passwordCheck.js"></script>
     <script type="module" src="signup.js"></script>
-    <!-- TODO password.js 나중에 합치기 -->
   </body>
 </html>
